order api implemented

+++ src/main/java/com/example/demo/Dto/OrderDto.java
@@ -20,12 +20,9 @@ public class OrderDto {
 
     private int amount;
 
-
-    private int price;
-
+    private Double price;
 
     private int orderId;
-
 
     private int goodId;
 
@@ -42,11 +39,11 @@ public class OrderDto {
         this.amount = amount;
     }
 
-    public int getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
